Validate event on creation
This provides a stronger demonstration of a result handler with multiple
possible outcomes (each of which receives interesting information--the
created event in the happy path, and the list of validation errors in
the invalid case).

The EventsController was getting cluttered, so I separated the create
endpoint from the get endpoints.